add Edit feature,delete modal window,scroll into view & Store data in local storage

+++ src/App.jsx
@@ -1,11 +1,16 @@
-import { useState } from 'react';
+import { useState, useRef, useEffect } from 'react';
 import { EditContext } from './contexts/EditContext';
 import Comments from './components/Comments';
 import info from '../src/data/data.json';
 import SendMsg from './components/SendMsg';
 export default function App() {
-  const [data, setData] = useState(info);
-  console.log(data);
+  const pastData = JSON.parse(localStorage.getItem('chatHistory'));
+  const connectedData = pastData ? pastData : info;
+  const [data, setData] = useState(connectedData);
+  useEffect(() => {
+    localStorage.setItem('chatHistory', JSON.stringify(data));
+  }, [data]);
+
   function deleteByID(comments, matchID) {
     return comments
       .filter(comment => comment.id !== matchID)
@@ -19,13 +24,21 @@ export default function App() {
         return comment;
       });
   }
+  const secRef = useRef(null);
   function handleComment(info) {
     const deleteData = info.delete;
-    const newData = { ...data };
-    newData.comments = deleteByID(newData.comments, deleteData);
-    setData(newData);
+    if (deleteData) {
+      const newData = { ...data };
+      newData.comments = deleteByID(newData.comments, deleteData);
+      setData(newData);
+    }
   }
-
+  useEffect(() => {
+    secRef.current.scrollIntoView({
+      behavior: 'smooth',
+      block: 'end',
+    });
+  }, [data.comments.length]);
   return (
     <EditContext.Provider value={{ handleComment, data, setData }}>
       <main className="mainApp">
@@ -37,6 +50,7 @@ export default function App() {
           setData={setData}
         />
       </main>
+      <div className="lastScroll" ref={secRef}></div>
     </EditContext.Provider>
   );
 }
